+++ src/lib/forms-schema/elements/ComplianceElement.ts
@@ -12,6 +12,7 @@ import {
   lookupSchemas,
   defaultValueOptionsSingle,
   customCssClasses,
+  hintPosition,
 } from '../property-schemas'
 
 export const type = 'compliance'
@@ -23,6 +24,7 @@ export default Joi.object({
   readOnly,
   ...requiredSchemas,
   hint,
+  hintPosition,
   storageType,
   ...optionsSchemas,
   defaultValue: defaultValueOptionsSingle,