+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `RECEIPT_ID` option to external id generation for forms
 - `postSubmissionReceipt` property to form validation
+- `hintPosition` property to element validation
 
 ## [2.0.2] - 2023-03-13